requirements.txt
# Hallucination Detection Middleware v2.1 — Free Edition (Ollama)
# ============================================================

# Core
openai>=1.30.0          # Ollama's OpenAI-compatible API (FREE, no key)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Vector store
chromadb>=0.5.0

# Hybrid search (BM25)
rank-bm25>=0.2.2

# Cross-encoder re-ranking
sentence-transformers>=3.0.0

# Document ingestion
pypdf>=4.0.0
beautifulsoup4>=4.12.0

# Dashboard is React-based — streamlit/plotly removed

# Evaluation harness
pyyaml>=6.0.1

# Cache backend (optional — falls back to in-memory if not installed/connected)
redis>=5.0.0

# Environment
python-dotenv>=1.0.0

# CLI & display
rich>=13.7.0
typer>=0.12.0

# ---- NEW in v2.1 ----

# Wikipedia ingestion (free, no API key)
wikipedia-api>=0.6.0

# Live web search — Tavily (AI-optimized, needs free API key from tavily.com)
tavily-python>=0.3.0

# Live web search — DuckDuckGo fallback (completely free, no API key)
ddgs>=9.0.0

# NLP — SpaCy entity extraction
spacy>=3.7.0
# After installing: python -m spacy download en_core_web_sm

# Optional: Anthropic (only needed if you switch LLM_PROVIDER=anthropic in .env)
# anthropic>=0.40.0

# ---- GPU acceleration (auto-detected at runtime) ----
# sentence-transformers and ChromaDB embeddings use GPU when torch+CUDA available.
# Remove this line if you have no NVIDIA GPU — CPU fallback is automatic.
torch>=2.1.0

# ---- Testing ----
pytest>=8.0.0
pytest-asyncio>=0.23.0
